Major refactor and first pass at new API
