+++ dagbldr/__init__.py
@@ -1,6 +1,7 @@
 __version_info__ = ('0', '0', '1')
 __version__ = '.'.join(__version_info__)
 
+from .core import *
 from .datasets import *
 from .nodes import *
 from .utils import *
