+++ dagbldr/core/__init__.py
@@ -1,6 +1,4 @@
 __version_info__ = ('0', '0', '1')
 __version__ = '.'.join(__version_info__)
 
-from ..core import *
-from .utils import *
-from .test_utils import *
+from .core import *
