+++ dagbldr/nodes/__init__.py
@@ -1,5 +1,5 @@
 __version_info__ = ('0', '0', '1')
 __version__ = '.'.join(__version_info__)
 
-from .layers import *
+from .nodes import *
 from .penalties import *
