+++ dagbldr/tests/test_full_feedforward.py
@@ -1,38 +1,40 @@
-from collections import OrderedDict
 import numpy as np
 import theano
+from theano import tensor
 
 from dagbldr.datasets import minibatch_iterator
-from dagbldr.utils import add_datasets_to_graph, convert_to_one_hot
-from dagbldr.utils import get_params_and_grads
-from dagbldr.utils import TrainingLoop
-from dagbldr.nodes import linear_layer, softmax_layer
+from dagbldr.utils import convert_to_one_hot
+from dagbldr import get_params, del_shared
+from dagbldr.training import TrainingLoop
+from dagbldr.nodes import linear, softmax
+
 from dagbldr.nodes import categorical_crossentropy
 from dagbldr.optimizers import sgd
 from dagbldr.datasets import load_digits
 
 # Common between tests
 digits = load_digits()
-X = digits["data"]
+X = digits["data"].astype("float32")
 y = digits["target"]
 n_classes = len(set(y))
-y = convert_to_one_hot(y, n_classes)
+y = convert_to_one_hot(y, n_classes).astype("float32")
 
 
 def test_feedforward_classifier():
+    del_shared()
     minibatch_size = 100
     random_state = np.random.RandomState(1999)
-    graph = OrderedDict()
+    X_sym = tensor.fmatrix()
+    y_sym = tensor.fmatrix()
 
-    X_sym, y_sym = add_datasets_to_graph([X, y], ["X", "y"], graph)
-
-    l1_o = linear_layer([X_sym], graph, 'l1', proj_dim=20,
-                        random_state=random_state)
-    y_pred = softmax_layer([l1_o], graph, 'pred', n_classes,
-                           random_state=random_state)
+    l1_o = linear([X_sym], [X.shape[1]], proj_dim=20, name='l1',
+                  random_state=random_state)
+    y_pred = softmax([l1_o], [20], proj_dim=n_classes, name='out',
+                     random_state=random_state)
 
     cost = categorical_crossentropy(y_pred, y_sym).mean()
-    params, grads = get_params_and_grads(graph, cost)
+    params = list(get_params().values())
+    grads = theano.grad(cost, params)
     learning_rate = 0.001
     opt = sgd(params, learning_rate)
     updates = opt.updates(params, grads)
@@ -56,20 +58,21 @@ def test_feedforward_classifier():
 
 
 def test_feedforward_theano_mix():
+    del_shared()
     minibatch_size = 100
     random_state = np.random.RandomState(1999)
-    graph = OrderedDict()
+    X_sym = tensor.fmatrix()
+    y_sym = tensor.fmatrix()
 
-    X_sym, y_sym = add_datasets_to_graph([X, y], ["X", "y"], graph)
-
-    l1_o = linear_layer([X_sym], graph, 'l1', proj_dim=20,
-                        random_state=random_state)
+    l1_o = linear([X_sym], [X.shape[1]], proj_dim=20, name='l1',
+                  random_state=random_state)
     l1_o = .999 * l1_o
-    y_pred = softmax_layer([l1_o], graph, 'pred', n_classes,
-                           random_state=random_state)
+    y_pred = softmax([l1_o], [20], proj_dim=n_classes, name='out',
+                     random_state=random_state)
 
     cost = categorical_crossentropy(y_pred, y_sym).mean()
-    params, grads = get_params_and_grads(graph, cost)
+    params = list(get_params().values())
+    grads = theano.grad(cost, params)
     learning_rate = 0.001
     opt = sgd(params, learning_rate)
     updates = opt.updates(params, grads)
