+++ examples/deprecated/faces_conditional_vae/flying_conditional_faces.py
