+++ examples/deprecated/faces_vae/faces_vae.py
