+++ examples/deprecated/faces_vae/flying_faces_vae.py
