+++ examples/deprecated/fruitspeech_ctc/fruitspeech_ctc.py
