+++ examples/deprecated/mixture_density_network/mixture_density_network.py
