+++ examples/deprecated/mnist_classifier/bn_mlp.py
