+++ examples/deprecated/mnist_classifier/bn_relu_mlp.py
