+++ examples/deprecated/mnist_classifier/logistic_sgd.py
