+++ examples/deprecated/mnist_classifier/mlp.py
