+++ examples/deprecated/mnist_conditional_vae/flying_mnist_conditional_vae.py
