+++ examples/deprecated/mnist_conditional_vae/mnist_conditional_vae.py
