+++ examples/deprecated/mnist_convolutional_classifier/convnet.py
