+++ examples/deprecated/mnist_vae/flying_mnist_vae.py
