+++ examples/deprecated/mnist_vae/mnist_vae.py
