+++ examples/deprecated/sinusoid_regression/sinusoid_regression.py
