+++ examples/deprecated/unconditional_handwriting_generation/sample_from_saved_handwriting_model.py
