+++ examples/deprecated/unconditional_handwriting_generation/unconditional_handwriting_generation.py
